refactor(frontend): separate institution admin entry
